requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
google-generativeai
aiofiles
scikit-learn
